requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
fastapi-mcp
openai==1.3.0
python-dotenv==1.0.0
